+++ src/confseq/misc.py
